+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed 
 - `linewidth` parameter added to `plot` method
 - `linewidth` parameter added to `plot_params_filter` function
 ## [0.5] - 2022-03-21