+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `save_params_filter` function
 ### Changed
-- `depth` parameter added to GenerativeImage `__init__`
+- `depth` section added to config/data file
 - `linewidth` parameter added to `plot` method
 - `linewidth` parameter added to `plot_params_filter` function
 ## [0.5] - 2022-03-21