+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `save_params_filter` function
 ### Changed
+- `link` field added to the `nft_storage_upload` result json
 - `depth` section added to config/data file
 - `linewidth` parameter added to `plot` method
 - `linewidth` parameter added to `plot_params_filter` function