+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `save_params_filter` function
 ### Changed
+- `message` field changed in `save_fig_file`
+- `message` field changed in `save_config_file`
+- `message` field changed in `save_data_file`
+- `message` field changed in `nft_storage_upload`
 - `depth` section added to config/data file
 - `linewidth` parameter added to `plot` method
 - `linewidth` parameter added to `plot_params_filter` function